backend/requirements.txt
google-adk>=0.1.0
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
google-cloud-firestore>=2.13.0
google-cloud-aiplatform>=1.38.0
firebase-admin>=6.2.0
python-dotenv>=1.0.0
pydantic>=2.4.0
google-generativeai>=0.3.0
python-multipart>=0.0.6
google-cloud-storage>=2.10.0
